alfa/requirements.txt
Flask
face_recognition
numpy
Pillow
opencv-python
qrcode[pil]
